[ADD] pxgo_bank_statement_analytic: New module to create analytic entries from bank statements.
Extends the Bank Statements to add support for analytic accounting.

A analytic account field will be added to bank statement lines, allowing
the user to directly register small expenses or incomes on analytic accounts.

This may be useful for petty cash management.

Thanks to Igalia (http://www.igalia.com/) for sponsoring this module :)